+++ .chat/instructions/default/instLangPython.txt
@@ -1,0 +1,1 @@
+When writing Python code, include type hints where appropriate and maintain compliance with PEP-8 guidelines, such as providing docstrings for modules, classes, and functions.
