+++ package.json
@@ -2,7 +2,7 @@
   "name": "devchat",
   "displayName": "devchat",
   "description": "devchat",
-  "version": "0.0.1",
+  "version": "0.0.3",
   "engines": {
     "vscode": "^1.77.0"
   },
@@ -24,7 +24,9 @@
         "DevChat.llmModel": {
           "type": "string",
           "default": "OpenAI",
-          "enum": ["OpenAI"],
+          "enum": [
+            "OpenAI"
+          ],
           "description": "Select whose llm to use."
         },
         "DevChat.OpenAI.model": {
@@ -123,7 +125,6 @@
     "@babel/preset-typescript": "^7.21.5",
     "@types/glob": "^8.1.0",
     "@types/jest": "^29.5.1",
-    "@types/mocha": "^10.0.1",
     "@types/ncp": "^2.0.5",
     "@types/node": "16.x",
     "@types/react-dom": "^18.2.3",
@@ -142,7 +143,6 @@
     "html-webpack-plugin": "^5.5.1",
     "jest": "^29.5.0",
     "json-loader": "^0.5.7",
-    "mocha": "^10.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
