+++ src/context/contextGitDiffCached.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import { ChatContext } from './contextManager';
-import { createTempSubdirectory, runCommandAndWriteOutput } from './commonUtil';
+import { createTempSubdirectory, runCommandAndWriteOutput } from '../util/commonUtil';
 
 export const gitDiffCachedContext: ChatContext = {
   name: 'git diff cached',
