+++ src/handler/doCommit.ts
@@ -1,0 +1,19 @@
+import * as vscode from 'vscode';
+import DtmWrapper from '../toolwrapper/dtm';
+import {messageHandler} from './messageHandler';
+
+
+async function doCommit(message: any, panel: vscode.WebviewPanel): Promise<void> {
+	const dtmWrapper = new DtmWrapper();
+
+	const commitResult = await dtmWrapper.commit(message.content);
+	if (commitResult.status === 0) {
+		vscode.window.showInformationMessage('Commit successfully.');
+	} else {
+		vscode.window.showErrorMessage(`Error commit fail: ${commitResult.message} ${commitResult.log}`);
+	}
+	return;
+}
+
+messageHandler.registerHandler('doCommit', doCommit);
+
