+++ src/util/commonUtil.ts
